requirements.txt
Flask==1.1.1
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
click==7.0
gunicorn==19.6.0
itsdangerous==0.24
beautifulsoup4
nltk==3.4.5
Flask-SSLify==0.1.5
epitran==1.8
python-dotenv==0.20.0
requests==2.27.1
tweepy==4.8.0
